[backend] rename compiler to codegen
